Commit work in progress before the weekend.
This isn't currently working, but the idea is that the config and management is all done through the calico script.

On coreos where python isn't installed, a binary can be created with create_binary.sh
A virtual env can be used when working with calico.py but my requirements.txt might not be complete.